changed the circle filter

+++ main.c
@@ -9,20 +9,20 @@
 #include <stdint.h>
 
 
-volatile uint16_t pointer[Num_of_Results];
-volatile uint16_t middle[Num_of_Results];
-volatile uint16_t ring[Num_of_Results];
-volatile uint16_t pinky[Num_of_Results];
+volatile unsigned int pointer[Num_of_Results];
+volatile unsigned int middle[Num_of_Results];
+volatile unsigned int ring[Num_of_Results];
+volatile unsigned int pinky[Num_of_Results];
 
-volatile uint16_t pointersum;
-volatile uint16_t middlesum;
-volatile uint16_t ringsum;
-volatile uint16_t pinkysum;
+volatile unsigned int pointersum;
+volatile unsigned int middlesum;
+volatile unsigned int ringsum;
+volatile unsigned int pinkysum;
 
-volatile uint16_t pointerval;
-volatile uint16_t middleval;
-volatile uint16_t ringval;
-volatile uint16_t pinkyval;
+volatile unsigned int pointerval;
+volatile unsigned int middleval;
+volatile unsigned int ringval;
+volatile unsigned int pinkyval;
 
 void main(void)
 {
@@ -34,6 +34,7 @@ void main(void)
     while(1){
     	 ADC14->CTL0 |= ADC14_CTL0_ENC |
     	     ADC14_CTL0_SC;				// Enable the adc and start conversion
+    	 //printf(" %d \n", pointerval);
     }
